% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{hormaleR}
\alias{hormaleR}
\title{Grabs the 50 most recent tweets from Male Republicans in the House of Representatives
Saves those tweets in individual csv (for each MC) and uses the UTF-8 file encoding for the CSV
Creates one dataframe to combine all of the tweets from each MC
With the one large dataframe, it creates a CSV to hold the 50 most recent tweets for each Male Republican in the House of Representatives}
\usage{
hormaleR()
}
\description{
Grabs the 50 most recent tweets from Male Republicans in the House of Representatives
Saves those tweets in individual csv (for each MC) and uses the UTF-8 file encoding for the CSV
Creates one dataframe to combine all of the tweets from each MC
With the one large dataframe, it creates a CSV to hold the 50 most recent tweets for each Male Republican in the House of Representatives
}
