% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{senfemR}
\alias{senfemR}
\title{Grabs the 50 most recent tweets from Female Republicans in the Senate
Saves those tweets in individual csv (for each MC) and uses the UTF-8 file encoding for the CSV
Creates one dataframe to combine all of the tweets from each MC
With the one large dataframe, it creates a CSV to hold the 50 most recent tweets for each Female Republican in the Senate}
\usage{
senfemR()
}
\description{
Grabs the 50 most recent tweets from Female Republicans in the Senate
Saves those tweets in individual csv (for each MC) and uses the UTF-8 file encoding for the CSV
Creates one dataframe to combine all of the tweets from each MC
With the one large dataframe, it creates a CSV to hold the 50 most recent tweets for each Female Republican in the Senate
}
