% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{token}
\alias{token}
\title{RStudio Console prompts the user to enter their twitter credentials
Users should keep these keys handy when starting a new session of R and should run this function before the others.}
\usage{
token()
}
\description{
RStudio Console prompts the user to enter their twitter credentials
Users should keep these keys handy when starting a new session of R and should run this function before the others.
}
