% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet_per_line.R
\name{tweet_per_line}
\alias{tweet_per_line}
\title{Takes each dataframe for the tweets that you downloaded and removes unimportant information if you are only interested in who posted it and the text of  the tweet
Writes a txt file with each tweet per row}
\usage{
tweet_per_line()
}
\description{
Takes each dataframe for the tweets that you downloaded and removes unimportant information if you are only interested in who posted it and the text of  the tweet
Writes a txt file with each tweet per row
}
